+++ src/classes/babies/PunkboyBaby.ts
@@ -2,7 +2,7 @@ import { CollectibleType } from "isaac-typescript-definitions";
 import { addCollectible, removeCollectible, repeat } from "isaacscript-common";
 import { Baby } from "../Baby";
 
-/** Starts with N Knife Piece 1 + N Knife Piece 2. */
+/** Starts with Nx Knife Piece 1 + Nx Knife Piece 2. */
 export class PunkboyBaby extends Baby {
   override onAdd(player: EntityPlayer): void {
     const num = this.getAttribute("num");